Adicionando login, arrumando comida e livros

+++ javali/src/main/java/javali/Controle/ControleComida.java
@@ -14,7 +14,7 @@ public class ControleComida {
         dao = new ComidaDAO();
     }
     
-    public void controleCadastrarComida(String nome, String descricao, int quantidade, double preco)
+    public void controleCadastrarComida(String nome, String descricao, int quantidade, float preco)
     throws IOException, SQLException, ClassNotFoundException{
       Comida comida = new Comida(nome, descricao, quantidade, preco);
       dao.cadastrarComidaDAO(comida);
